# backend/internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/authorization": {
            "get": {
                "tags": [
                    "authorization"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page_index",
                        "name": "page_index",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page_size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "node id",
                        "name": "node_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "asset id",
                        "name": "asset_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "account id",
                        "name": "account_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.HttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/controller.ListData"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/model.Account"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "authorization"
                ],
                "parameters": [
                    {
                        "description": "authorization",
                        "name": "authorization",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Authorization"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.HttpResponse"
                        }
                    }
                }
            }
        },
        "/authorization/:id": {
            "delete": {
                "tags": [
                    "authorization"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "authorization id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.HttpResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.HttpResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "controller.ListData": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "list": {
                    "type": "array",
                    "items": {}
                }
            }
        },
...
        "model.Account": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "account_type": {
                    "type": "integer"
                },
                "asset_count": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "phrase": {
                    "type": "string"
                },
                "pk": {
                    "type": "string"
                },
                "resource_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "updater_id": {
                    "type": "integer"
                }
            }
        },
...
        "model.Authorization": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "integer"
                },
                "asset_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "node_id": {
                    "type": "integer"
                },
                "resource_id": {
                    "type": "integer"
                },
                "rids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "updater_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}